# api-specs/api-spec-v3.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/images':
    get:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Get an image
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '201':
          description: Image created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageId:
                    type: integer
                    description: The ID of the image that has just been created
                    example: 1234
                  fileName:
                    type: string
                    description: The URL of the image that has just been created
                    example: media/images/1234.png
                  thumbnail_Filename:
                    type: string
                    description: The URL of the image thumbnail that has just been created
                    example: media/images/1234_thumbnail.png
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to add an image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
  '/businesses/{businessId}/products/{productId}/images':
    post:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Upload a new image for some photo
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business to search catalogue of
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to try to find in catalogue
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                filename:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Image created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageId:
                    type: integer
                    description: The ID of the image that has just been created
                    example: 1234
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to add an image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{businessId}/products/{productId}/images/{imageId}':
    delete:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Delete an image of a product
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to delete image for
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          description: ID of the image to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to delete an image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/businesses/{businessId}/products/{productId}/images/{imageId}/makeprimary':
    put:
      tags:
        - 'U16: Product Images'
        - 'Catalogue / Products'
      summary: Set an existing image to be the new primary image for some product
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: businessId
          in: path
          description: ID of business that owns product
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          description: ID (Product Code) of product to update primary image for
          required: true
          schema:
            type: string
        - name: imageId
          in: path
          description: ID of the image to set as being the primary image for product
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Primary image successfully updated
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 'Forbidden: Returned when a user tries to update the primary image for a product that it is in a catalogue for a business they do not administer AND the user is not a global application admin'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: There was some error with the data supplied by the user, appropriate error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: 'The requested route does exist (so not a 404) but some part of the request is not acceptable, for example trying to access a resource by an ID that does not exist.'
...